fix: the totalSize in page request

+++ src/main/java/com/movie/rating/interfaces/controller/representation/PageRepresentation.java
@@ -2,5 +2,5 @@ package com.movie.rating.interfaces.controller.representation;
 
 import java.util.List;
 
-public record PageRepresentation<T>(List<T> contents, Integer totalSize) {
+public record PageRepresentation<T>(List<T> contents, Long totalSize) {
 }
